Renaming common file to 'add or update exhibitor'
