+++ components/expofp/actions/add-exhibitor/add-exhibitor.mjs
@@ -1,4 +1,4 @@
-import common from "../common.mjs";
+import common from "../common/add-or-update-exhibitor.mjs";
 
 export default {
   ...common,
