+++ components/expofp/actions/update-exhibitor/update-exhibitor.mjs
@@ -1,5 +1,5 @@
 import expofp from "../../expofp.app.mjs";
-import common from "../common.mjs";
+import common from "../common/add-or-update-exhibitor.mjs";
 
 export default {
   ...common,
